Add the ability to watch for PrometheusCluster CRD events
This change adds the ability for the operator to continuously watch for
events relating to the PrometheusCluster CRD. Currently, it does not do
anything with these events other than log the event. You can check if
the operator is able to detect it by tailing the logs of the oeprator's
pod.