Fully replace bespoke code with helm calls to manage prometheus
This change removes all the bespoke Python code and instead creates a
wrapper for the Helm 3 CLI (included in the generated container image).
This gives the operator install, uninstall, and upgrade capabilities for
free with very little coding. Because of this, we might be able to add
blue/green deployment sooner than expected.